requirements.txt
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
deprecation==2.1.0
Flask==2.2.3
functions-framework==3.3.0
google-api-core==2.11.0
google-auth==2.17.2
google-cloud-core==2.3.2
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypdf==3.7.0
requests==2.28.2
rsa==4.9
six==1.16.0
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
